DeleteBuildingByIdNotFoundShouldThrowExceptionTestLogic Green

+++ BusinessLogic/BuildingLogic.cs
@@ -48,6 +48,10 @@ namespace BusinessLogic
                 throw new ArgumentException("Id is empty", nameof(id));
             }
             Building building = _buildingRepository.GetBuildingById(id);
+            if (building == null)
+            {
+                throw new ArgumentException("Building not found", nameof(id));
+            }
             _buildingRepository.DeleteBuilding(building);
             _buildingRepository.Save();
         }
